// src/shared/giftcard/smp-giftcard-transaction-ops.enum.ts
export enum SmpGiftcardTransactionOps {
  ACTIVATE = "ACTIVATE",
  BLOCK = "BLOCK",
  CHECK = "CHECK",
  DIGITAL_ACTIVATE = "DIGITAL_ACTIVATE",
  FREEZE = "FREEZE",
  PURCHASE = "PURCHASE",
  RECHARGE = "RECHARGE",
  ROLLBACK = "ROLLBACK",
  TRANSACTION = "TRANSACTION",
  TRANSACTION_DETAIL = "TRANSACTION_DETAIL",
  UNFREEZE = "UNFREEZE",
  XREFUND = "XREFUND", // Workaround for partial refunds
}